# src/main/java/com/ascend/testlab/dto/response/ExperimentHistoryResponse.java
package com.ascend.testlab.dto.response;

import com.ascend.testlab.dto.entity.experiment.ExperimentHistoryEntry;
import java.util.List;
import lombok.Builder;

/**
 * Response DTO for the Get Experiment History API containing the experiment id, history entries,
 * and pagination metadata.
 *
 * @author Nithya sree
 * @version 1.0
 * @since 1.0
 */
@Builder
public record ExperimentHistoryResponse(
    String experiment_id, List<ExperimentHistoryEntry> history, PaginationMeta pagination) {}