# src/main/java/com/ascend/testlab/dto/response/ExperimentKeyAvailabilityResponse.java
package com.ascend.testlab.dto.response;

/**
 * Response DTO for the Experiment Key Availability API containing the availability status.
 *
 * @author Nithya sree
 * @version 1.0
 * @since 1.0
 */
public record ExperimentKeyAvailabilityResponse(Boolean is_available) {}