# src/main/resources/webroot/swagger/swagger.yaml
openapi: 3.0.1
info:
  title: testlab
  description: testlab Service
  version: "1.0"
paths:
  /v1/allocations:
    get:
      operationId: getAllocationHandle
      parameters:
      - name: x-project-key
        in: header
        schema:
          type: string
      - name: x-user-Id
        in: header
        schema:
          type: string
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        "304":
          description: Successful Not Modified Response
        "400":
          description: Bad Request due to invalid/missing body params / header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "401":
          description: Bad Request due to invalid project-key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
    put:
      operationId: reAllocateExperimentHandle
      parameters:
      - name: x-project-key
        in: header
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReallocateRequest'
        required: true
      responses:
        "200":
          description: Successfully reallocated user to new variant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        "400":
          description: Bad Request - invalid/missing body params or header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "404":
          description: Not Found - Experiment or user assignment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
    post:
      operationId: allocationHandle
      parameters:
      - name: x-project-key
        in: header
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AllocationRequest'
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        "304":
          description: Successful Not Modified Response
        "400":
          description: Bad Request due to invalid/missing body params / header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "401":
          description: Bad Request due to invalid project-key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
  /v1/experiments/{experiment_id}:
    get:
      operationId: getExperimentHandle
      parameters:
      - name: x-project-key
        in: header
        required: true
        schema:
          type: string
      - name: experiment_id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        "400":
          description: Experiment Id is not valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "404":
          description: Experiment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
    delete:
      operationId: handle
      parameters:
      - name: x-project-key
        in: header
        required: true
        schema:
          type: string
      - name: experiment_id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        "400":
          description: Project id is missing or Experiment id is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "404":
          description: Experiment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
  /v1/experiments/key-availability:
    get:
      operationId: experimentKeyAvailabilityHandle
      parameters:
      - name: x-project-key
        in: header
        required: true
        schema:
          type: string
      - name: experiment_key
        in: query
        required: true
        schema:
          maxLength: 255
          minLength: 0
          type: string
      responses:
        "200":
          description: Successfully validated experiment key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessExperimentKeyAvailabilityResponse'
        "400":
          description: Invalid project key or experiment key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
  /v1/experiments:
    get:
      operationId: filterExperimentsHandle
      parameters:
      - name: x-project-key
        in: header
        required: true
        schema:
          type: string
      - name: status
        in: query
        schema:
          type: string
      - name: owner
        in: query
        schema:
          type: string
      - name: name
        in: query
        schema:
          type: string
      - name: type
        in: query
        schema:
          type: string
      - name: tag
        in: query
        schema:
          type: string
      - name: limit
        in: query
        schema:
          type: integer
          format: int32
          default: 20
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 1
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        "400":
          description: Project Key is missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
  /v1/experiments/{experiment_id}/history:
    get:
      operationId: getExperimentHistoryHandle
      parameters:
      - name: x-project-key
        in: header
        required: true
        schema:
          type: string
      - name: experiment_id
        in: path
        required: true
        schema:
          type: string
      - name: limit
        in: query
        schema:
          type: integer
          format: int32
          default: 20
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 1
      responses:
        "200":
          description: Successfully fetched experiment history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessExperimentHistoryResponse'
        "400":
          description: Invalid project key or experiment id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "404":
          description: Experiment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
  /v1/experiments/tags:
    get:
      operationId: getTagsHandle
      parameters:
      - name: x-project-key
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successfully retrieved tags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessTagsResponse'
        "400":
          description: Invalid project Key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
components:
  schemas:
    Success:
      type: object
      properties:
        data:
          type: object
    ErrorEntity:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        cause:
          type: string
    Failure:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorEntity'
    AllocationResponse:
      type: object
      properties:
        experiment_map:
          type: array
          items:
            $ref: '#/components/schemas/UserExperimentMap'
    SuccessAllocationResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/AllocationResponse'
    UserExperimentMap:
      type: object
      properties:
        status:
          type: string
        variant:
          $ref: '#/components/schemas/Variant'
        experiment_id:
          type: string
          format: uuid
        experiment_name:
          type: string
        variant_name:
          type: string
        assigned_at:
          type: integer
          format: int64
    Variables:
      required:
      - dataType
      type: object
      properties:
        value:
          type: string
        key:
          type: string
        dataType:
          type: string
    Variant:
      type: object
      properties:
        variables:
          type: array
          items:
            $ref: '#/components/schemas/Variables'
        display_name:
          type: string
    AllocationRequest:
      type: object
      properties:
        attributes:
          $ref: '#/components/schemas/Attributes'
        experiment_keys:
          type: array
          items:
            type: string
        stable_id:
          type: string
        user_id:
          type: string
    Attributes:
      type: object
      properties:
        model:
          type: string
        device:
          type: string
        platform:
          type: string
        build_version:
          type: string
        app_name:
          type: string
        os_version:
          type: string
        app_version:
          type: string
        build_number:
          type: string
    GetAllocationsResponse:
      type: object
      properties:
        experiment_map:
          type: array
          items:
            $ref: '#/components/schemas/UserExperimentMap'
    SuccessGetAllocationsResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/GetAllocationsResponse'
    SuccessUserExperimentMap:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/UserExperimentMap'
    ReallocateRequest:
      required:
      - experiment_id
      - user_id
      - variant_name
      type: object
      properties:
        reason:
          type: string
        experiment_id:
          type: string
        variant_name:
          type: string
        user_id:
          type: string
    SuccessBoolean:
      type: object
      properties:
        data:
          type: boolean
    ExperimentKeyAvailabilityResponse:
      type: object
      properties:
        is_available:
          type: boolean
    SuccessExperimentKeyAvailabilityResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ExperimentKeyAvailabilityResponse'
    CohortVariantWeights:
      type: object
      allOf:
      - $ref: '#/components/schemas/VariantWeights'
      - type: object
        properties:
          weights:
            type: object
            additionalProperties:
              type: number
              format: double
    Condition:
      required:
      - operand
      - operandDataType
      - operator
      - value
      type: object
      properties:
        operand:
          type: string
        operandDataType:
          type: string
        operator:
          type: string
        value:
          type: string
    Experiment:
      type: object
      properties:
        projectKey:
          type: string
        name:
          type: string
        key:
          type: string
        description:
          type: string
        hypothesis:
          type: string
        status:
          type: string
          enum:
          - LIVE
          - PAUSED
          - DRAFT
          - CONCLUDED
          - TERMINATED
        type:
          type: string
          enum:
          - A/A
          - A/B
        guardrailHealthStatus:
          type: string
          enum:
          - WARNING
          - PASSED
          - NO_CHECKS_AVAILABLE
          - FAILED
        cohorts:
          type: array
          items:
            type: string
        variantWeights:
          $ref: '#/components/schemas/VariantWeights'
        variants:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Variant'
        ruleAttributes:
          type: array
          items:
            $ref: '#/components/schemas/RuleAttributes'
        distributionStrategy:
          type: string
          enum:
          - ROUND_ROBIN
          - RANDOM
        assignmentDomain:
          type: string
          enum:
          - STRATIFIED
          - COHORT
        overrides:
          type: array
          items:
            type: string
        winningVariant:
          $ref: '#/components/schemas/JsonObject'
        exposure:
          type: integer
          format: int32
        threshold:
          type: integer
          format: int64
        startTime:
          type: integer
          format: int64
        endTime:
          type: integer
          format: int64
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
        owners:
          type: array
          items:
            type: string
        experiment_id:
          type: string
          format: uuid
    FilterExperimentsResponse:
      type: object
      properties:
        experiments:
          type: array
          items:
            $ref: '#/components/schemas/Experiment'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    JsonObject:
      type: object
      properties:
        map:
          type: object
          additionalProperties:
            type: object
        empty:
          type: boolean
    PaginationMeta:
      type: object
      properties:
        current_page:
          type: integer
          format: int32
        page_size:
          type: integer
          format: int32
        total_count:
          type: integer
          format: int32
        has_next:
          type: boolean
    RuleAttributes:
      required:
      - name
      type: object
      properties:
        name:
          type: string
        conditions:
          type: array
          items:
            $ref: '#/components/schemas/Condition'
    StratifiedVariantWeights:
      type: object
      allOf:
      - $ref: '#/components/schemas/VariantWeights'
      - type: object
        properties:
          weights:
            type: object
            additionalProperties:
              type: array
              items:
                type: string
    SuccessFilterExperimentsResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/FilterExperimentsResponse'
    VariantWeights:
      type: object
      properties:
        type:
          type: string
          enum:
          - STRATIFIED
          - COHORT
      discriminator:
        propertyName: type
    SuccessExperiment:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Experiment'
    ExperimentHistoryEntry:
      type: object
      properties:
        updatedBy:
          type: string
        previousData:
          $ref: '#/components/schemas/JsonObject'
        currentData:
          $ref: '#/components/schemas/JsonObject'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ExperimentHistoryResponse:
      type: object
      properties:
        experiment_id:
          type: string
        history:
          type: array
          items:
            $ref: '#/components/schemas/ExperimentHistoryEntry'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    SuccessExperimentHistoryResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ExperimentHistoryResponse'
    SuccessTagsResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/TagsResponse'
    TagsResponse:
      type: object
      properties:
        tags:
          type: array
          items:
            type: string